requirements.txt
face_recognition_models
Flask~=2.0.2